% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_feature_data_db.R
\name{collect_feature_data_db}
\alias{collect_feature_data_db}
\title{Collect Feature data after image analysis with cellprofiler}
\usage{
collect_feature_data_db(
  path_data,
  host_db = "biosensor.c9k2hfiwt5mi.us-east-2.rds.amazonaws.com",
  result_path = "~/dcp_helper/data/results/",
  channel_of_interest = "ch1",
  measurement_of_interest = "measurement_IdentifySecondaryObjects.csv"
)
}
\arguments{
\item{result_path}{Path to collected measurements}

\item{measurement_of_interest}{}
}
\value{

}
\description{
Collect Feature data after image analysis with cellprofiler
}
