% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_flatfield_metadata_split.R
\name{create_flatfield_metadata_split}
\alias{create_flatfield_metadata_split}
\title{Title}
\usage{
create_flatfield_metadata_split(
  path = paste0(getwd(), "/"),
  channel_of_interest,
  name = "bf",
  path_base = "~/bucket/metadata/",
  flatfield_dir = "flatfield",
  force = FALSE,
  include_brightfield_proj = FALSE,
  include_additional_proj = FALSE
)
}
\arguments{
\item{name}{}
}
\description{
Title
}
