% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_filelist.R
\name{extract_filelist}
\alias{extract_filelist}
\title{Return tibble describing each image file with well, field etc. in a given directory}
\usage{
extract_filelist(
  path = getwd(),
  force,
  path_base,
  path_yml = "~/mcsaba/biosensor/src/dcp_helper/python/pe2loaddata_config.yml"
)
}
\arguments{
\item{path}{the path of the directory to screen. The path is not checked recursively}
}
\value{
a tibble with columns: file_name, n_zst, well, col, fld, zst, l_row, channel, file_path, name, ext, parent, timepoint
}
\description{
Return tibble describing each image file with well, field etc. in a given directory
}
