% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_filelist_alternative.R
\name{extract_filelist_alternative}
\alias{extract_filelist_alternative}
\title{Return tibble describing each image file with well, field etc. in a given directory}
\usage{
extract_filelist_alternative(path = getwd(), force, path_base, path_yml)
}
\arguments{
\item{path}{the path of the directory to screen. The path is not checked recursively}
}
\value{
a tibble with columns: file_name, n_zst, well, col, fld, zst, l_row, channel, file_path, name, ext, parent, timepoint
}
\description{
Return tibble describing each image file with well, field etc. in a given directory
}
