% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_output_structure.R
\name{format_output_structure}
\alias{format_output_structure}
\title{format_output_structure}
\usage{
format_output_structure(metadata_tags)
}
\arguments{
\item{metadata_tags}{List of metadata values}
}
\value{
Formatted string for output structure
}
\description{
format_output_structure
}
