% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_data_file.R
\name{link_data_file}
\alias{link_data_file}
\title{Fill a JSON template with a specific metadata path and metadata groupings}
\usage{
link_data_file(
  metadata_split_path,
  json_template,
  json_path,
  path_base,
  flatfield_dir = "flatfield",
  path_to_metadata = "dcp_helper/metadata"
)
}
\arguments{
\item{metadata_split_path}{Table that contains metadata and file information for CellProfiler pipeline using LoadData module}

\item{path_base}{Destination directory of job files}

\item{flatfield_dir}{Directory for intemediate and final results}

\item{path_to_metadata}{Directory containing metadata files}
}
\description{
Fill a JSON template with a specific metadata path and metadata groupings
}
